refactor visualization.
