+++ example_scripts/plot_llm_capability_scores.py
@@ -7,10 +7,12 @@ from tqdm import tqdm
 
 from src.generate_capabilities import (
     get_previous_capabilities,
-    plot_capability_scores_spider_and_bar_chart,
     select_complete_capabilities,
 )
 from src.utils.data_utils import get_run_id
+from src.utils.visualization_utils import (
+    plot_capability_scores_spider_and_bar_chart,
+)
 
 
 logger = logging.getLogger(__name__)
